Filters to restrict granting/revoking access to files Closes #4952

+++ includes/admin/post-types/class-wc-admin-cpt-product.php
@@ -1069,15 +1069,19 @@ class WC_Admin_CPT_Product extends WC_Admin_CPT {
 					// Remove permissions
 					if ( $removed_download_ids ) {
 						foreach ( $removed_download_ids as $download_id ) {
-							$wpdb->query( $wpdb->prepare( "DELETE FROM {$wpdb->prefix}woocommerce_downloadable_product_permissions WHERE order_id = %d AND product_id = %d AND download_id = %s", $order->id, $product_id, $download_id ) );
+							if ( apply_filters( 'woocommerce_process_product_file_download_paths_remove_access_to_old_file', true, $download_id, $product_id, $order ) ) {
+								$wpdb->query( $wpdb->prepare( "DELETE FROM {$wpdb->prefix}woocommerce_downloadable_product_permissions WHERE order_id = %d AND product_id = %d AND download_id = %s", $order->id, $product_id, $download_id ) );
+							}
 						}
 					}
 					// Add permissions
 					if ( $new_download_ids ) {
 						foreach ( $new_download_ids as $download_id ) {
-							// grant permission if it doesn't already exist
-							if ( ! $wpdb->get_var( $wpdb->prepare( "SELECT 1 FROM {$wpdb->prefix}woocommerce_downloadable_product_permissions WHERE order_id = %d AND product_id = %d AND download_id = %s", $order->id, $product_id, $download_id ) ) ) {
-								wc_downloadable_file_permission( $download_id, $product_id, $order );
+							if ( apply_filters( 'woocommerce_process_product_file_download_paths_grant_access_to_new_file', true, $download_id, $product_id, $order ) ) {
+								// grant permission if it doesn't already exist
+								if ( ! $wpdb->get_var( $wpdb->prepare( "SELECT 1 FROM {$wpdb->prefix}woocommerce_downloadable_product_permissions WHERE order_id = %d AND product_id = %d AND download_id = %s", $order->id, $product_id, $download_id ) ) ) {
+									wc_downloadable_file_permission( $download_id, $product_id, $order );
+								}
 							}
 						}
 					}
